flood_alert_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator